backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.12.0
httpx==0.24.1
google-generativeai
opencv-python==4.9.0.80
mediapipe==0.10.21
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
PyPDF2==3.0.1
